Task/requirements.txt
mysql-connector-python >= 8